+++ src/Test Projects/TestWinFormDotNet45/Program.cs
@@ -46,10 +46,23 @@ namespace TestWinFormDotNet45
             //ExceptionHandler.Filter.JustRaiseErrorCodeScopes.Add(
             //    new CodeScope("Assembly", "Namespace", "Class", "Method"));
             //
+            // Show unhandled exception message customized mode. 
+            ExceptionHandler.OnShowUnhandledError += AlertUnhandledErrors;
+            //
             // ---------------------------------------------------------------------------------
             //
 
             Application.Run(new FormTest());
+        }
+
+        /// <summary>
+        /// Show unhandled exception message customized mode.
+        /// </summary>
+        /// <param name="sender">Raw exception object</param>
+        /// <param name="e">Compiled error object</param>
+        public static void AlertUnhandledErrors(object sender, UnhandledErrorEventArgs e)
+        {
+            MessageBox.Show(e.ErrorObject.Message);
         }
 
         public static void Exp()